hackonweb.client/backend-mlmodel/requirements.txt
Flask
flask-cors
python-pptx
transformers
torch